Add blogNav
